proto - battery and settings icon placeholder
Something is wrong with icons here, and I think it's probably something
in my flake? All of the expected icon names are showing up in IconTheme,
when programmatically queried, but `GTK_DEBUG=iconfallback` is emitting
"No icon found in <theme> (or fallbacks)" error messages. Beginning to
suspect the error message is just wrong for the actual failure, and the
error is actually something when trying to find icon data and get the
pixbuf? Don't know for sure though. Needs some digging.