styles - tranparentized themed colors pattern
Opacity is implemented as a ratio of the parent element's opacity, so
the child of a node with 60% opacity can never be fully opaque.

Working around this by using `color-mix` to mix themed colors with
`transparent` in different ratios, so that children can be more opaque
than their parents while still following GTK theme colors defined with
`@define-color`.