VanityBrightness - library for updating device brightness
Currently only implements monitoring. There is no dbus based monitoring,
so brightness is monitored via files in /sys/class/*. Updating
brightness incoming next, which does requrie dbus.

These are off in a library directory. I'll probably try to contribute
this upstream to astal, or might publish this as a library if not.