prototype for managing wlsunset via systemd dbus
I think I need to revisit this later. Method calls for start / stop /
restart are all working fine, but the bus proxy properties for the unit
are not ever being updated. I don't want to be repeatedly creating the
bus and polling on a timer, so I need to find some other signal to hook
into for updates. I think the right one might be JobNew and JobRemoved
in org.freedesktop.systemd1.Manager. Might come back to this after some
other higher priority work, or I might just use the old lockfile method
I used in my old bar if this is the only service using it.