dependencies - proto transition libglycin to nixpkgs-unstable package
This almost works. I'm not sure if there's something I'm missing or if
this is just a little broken pending upstream work. The sandboxing with
brwap tries to find the bwrap executable with `env -i "bwrap" ...`,
which obviously doesn't work on nixos because -i (ignore environment)
unsets path related variables and it can't find the bwrap executable.

Trying to find an answer, and going to try to contribute upstream if
there's still some work pending here.